+++ src/main/java/com/example/befacerecognitionattendance2025/service/AuthService.java
@@ -12,7 +12,6 @@ public interface AuthService {
 
     LoginResponse refreshToken(RefreshTokenRequest  request) ;
 
-    EmployeeResponse changePassword(ChangePasswordRequest request) ;
 
     EmployeeResponse verifyPassword(String password);
